Added mode to merge csv files

+++ ExcelMerger/Program.cs
@@ -1,32 +1,41 @@
 ﻿using OfficeOpenXml;
 using System.Text;
 
+namespace ExcelMerger;
+
 internal class Program
 {
 	static void Main(string[] args)
 	{
 		ExcelPackage.LicenseContext = LicenseContext.NonCommercial;
-		var names = GetStudentMap();
+		var names = GetStudentMap("std.csv");
+		// Check std merge mode
+		if (args.Length > 0 && args[0] == "-m")
+		{
+			MergeStudentMap(names);
+			return;
+		}
+
 		File.Delete("destination.xlsx");
 		using var source = new ExcelPackage(new FileInfo("raw.xlsx"));
 		using var destination = new ExcelPackage(new FileInfo("destination.xlsx"));
-		foreach (var sheet in source.Workbook.Worksheets)
+		foreach (ExcelWorksheet? sheet in source.Workbook.Worksheets)
 		{
 			// Add the sheet
-			var destinationWorksheet = destination.Workbook.Worksheets.Add(sheet.Name);
+			ExcelWorksheet? destinationWorksheet = destination.Workbook.Worksheets.Add(sheet.Name);
 			// Reverse columns if needed
 			if (args.Length > 0 && args[0] == "-r")
 				ReverseColumns(sheet);
 			// Copy first row
 			destinationWorksheet.Cells[1, 1].Value = "Name";
-			for (int column = 1; column <= sheet.Dimension.End.Column; column++)
+			for (var column = 1; column <= sheet.Dimension.End.Column; column++)
 				destinationWorksheet.Cells[1, column + 1].Value = Reverse(sheet.Cells[1, column].Text);
 			// Add names
-			for (int row = 2; row <= sheet.Dimension.End.Row; row++)
+			for (var row = 2; row <= sheet.Dimension.End.Row; row++)
 			{
 				string stdNumber = ToEnglishNumber(sheet.Cells[row, 1].Text);
 				destinationWorksheet.Cells[row, 1].Value = names.ContainsKey(stdNumber) ? names[stdNumber] : "ناشناس";
-				for (int column = 1; column <= sheet.Dimension.End.Column; column++)
+				for (var column = 1; column <= sheet.Dimension.End.Column; column++)
 				{
 					string data = ToEnglishNumber(sheet.Cells[row, column].Text);
 					if (double.TryParse(data, out double dataInt))
@@ -36,33 +45,52 @@ internal class Program
 				}
 			}
 		}
+
 		destination.Save();
 	}
 
-	private static Dictionary<string, string> GetStudentMap()
+	private static Dictionary<string, string> GetStudentMap(string path)
 	{
 		var names = new Dictionary<string, string>();
-		using var reader = new StreamReader("std.csv");
+		using var reader = new StreamReader(path);
 		reader.ReadLine(); // skip headers
-		string? line;
-		while((line = reader.ReadLine()) != null)
+		while (reader.ReadLine() is { } line)
 		{
 			string[] splitted = line.Split(',');
 			names.Add(splitted[0].Trim(), splitted[1].Trim());
 		}
+
 		return names;
+	}
+
+	private static void MergeStudentMap(IDictionary<string, string> original)
+	{
+		var newStudents = GetStudentMap("new-std.csv");
+		foreach ((string stdId, string stdName) in newStudents)
+			if (!original.ContainsKey(stdId))
+			{
+				original[stdId] = stdName;
+				Console.WriteLine($"Added {stdId}");
+			}
+
+		// Write back
+		using var writer = new StreamWriter("std.csv");
+		writer.WriteLine("stdId,stdName");
+		foreach ((string stdId, string stdName) in original)
+			writer.WriteLine($"{stdId},{stdName}");
 	}
 
 	private static string ToEnglishNumber(string input)
 	{
 		StringBuilder englishNumbers = new(input.Length);
-		for (int i = 0; i < input.Length; i++)
+		for (var i = 0; i < input.Length; i++)
 		{
 			if (char.IsDigit(input[i]))
 				englishNumbers.Append(char.GetNumericValue(input, i));
 			else
 				englishNumbers.Append(input[i]);
 		}
+
 		return englishNumbers.ToString();
 	}
